Add preconditions to AddValidToNestedConfigProperties
